refactor!: remove deprecations before the next release
