+++ .changeset/sixty-llamas-greet.md
@@ -1,0 +1,9 @@
+---
+"@shopware-pwa/helpers-next": major
+---
+
+Removed deprecations from the package:
+
+- `getProductThumbnailUrl` is removed. Use `getSmallestThumbnailUrl` instead.
+- internal method `_debounce` is removed.
+- internal method `_parseUrlQuery` is removed.
