+++ packages/helpers/src/index.ts
@@ -2,14 +2,14 @@ export * from "./product";
 export * from "./category";
 export * from "./ui-interfaces";
 export * from "./cms";
-export * from "./urlHelpers";
-export * from "./_debounce";
 export * from "./getTranslatedProperty";
 export * from "./listing";
 export * from "./price";
-export * from "./media/image";
-export * from "./media/getMedia";
 export * from "./media/downloadFile";
+export * from "./media/getBiggestThumbnailUrl";
+export * from "./media/getMedia";
+export * from "./media/getSmallestThumbnailUrl";
+export * from "./media/getSrcSetForMedia";
 export * from "./checkout";
 export * from "./languages";
 export * from "./relativeUrlSlash";
