+++ packages/helpers/src/product/index.ts
@@ -5,7 +5,6 @@ export * from "./getProductReviews";
 export * from "./getProductTierPrices";
 export * from "./getProductUrl";
 export * from "./getProductRatingAverage";
-export * from "./getProductThumbnailUrl";
 export * from "./getProductCalculatedListingPrice";
 export * from "./getProductFromPrice";
 export * from "./getProductRoute";
